+++ docs/Changelog.md
@@ -4,6 +4,8 @@
 
 - Upgraded to PowerAuth 1.9.x [(#165)](https://github.com/wultra/mtoken-sdk-ios/pull/165)
 - Document fixes and improvements
+- Added status to `UserOperation` and removed redundant `OperationHistoryEntry` [(#171)](https://github.com/wultra/mtoken-sdk-ios/pull/171)
+- Removed deprecated `IOperationsService` methods in 1.5.0. [(#171)](https://github.com/wultra/mtoken-sdk-ios/pull/171)
 
 ## 1.11.1 (Aug, 2024)